FLASH-134 refactor: UUIDGenerator 패키지 위치 이동
